support correct split on value converting strings
